Feature/tamba3 (#49)
* bug fix in PushToMongoDB
* do not exclude waveform telemetry
* daq_in_progress
* mongodb
* mt_test4.rb
* added error in MeasureTemperatureWithRTDSensor.cc
* do not allow negative voltage from DAC
* time stamp in ReceiveCommand.cc
* Added invalid command
* ShutdownSystem.cc
* overflow in TelemetryDefinition.cc
* bug fix in ControlHighVoltage.cc
* bug fix in ReceiveCommand.cc

---------

+++ onboard/source/core/src/ErrorManager.cc
@@ -11,7 +11,7 @@ ErrorManager::ErrorManager()
 
 void ErrorManager::resetError()
 {
-  errorCode_ = 0;
+  errorCode_ = (uint64_t)0;
 }
 
 void ErrorManager::setError(ErrorType v)
